requirements.txt
fathomnet
tqdm
requests
ultralytics
huggingface_hub
dill
pycocotools
opencv-python
pyyaml
numpy
